Refactor helper methods into BdvPlaygroundUtils
